Update documented rust version
Trying to deploy crate using Return Position Impl Trait In Trait, with rust-version set to 1.75, the deployer nicely prints that the currently active rustc version is 1.74.0. I noticed that was different than the documented behaviour. I have not checked that the environment variables are updated accordingly in the enviroment, but this commit assumes that.